+++ app/Http/Controllers/ItemPartialTransferManage.php
@@ -51,7 +51,7 @@ class ItemPartialTransferManage extends Controller
             if ($partial_transfer !== null) {
                 $partial_transfer->delete();
 
-                ClearCache::dispatchNow($cache_job_payload->payload());
+                ClearCache::dispatch($cache_job_payload->payload());
 
                 return Responses::successNoContent();
             }