+++ app/Http/Controllers/ItemSubcategoryManage.php
@@ -156,7 +156,7 @@ class ItemSubcategoryManage extends Controller
         try {
             $item_sub_category->delete();
 
-            ClearCache::dispatchNow($cache_job_payload->payload());
+            ClearCache::dispatch($cache_job_payload->payload());
 
             return \App\Response\Responses::successNoContent();
         } catch (QueryException $e) {